Accept single trigger in JSRule
The code block example suggests passing a single trigger instance is possible, but it had to be an array actually. This change now accepts both. I believe that's reasonable, as many rules will have just a single trigger (e.g. cron), so having to wrap in an array explicitly might be annoying.